requirements.txt
Django==4.2.16
oauthlib==3.2.2
psycopg==3.2.3
python-decouple==3.8
djangorestframework==3.15.2
django-cors-headers==4.5.0
